% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_table_SIR.R
\name{create_parameter_table_SIR}
\alias{create_parameter_table_SIR}
\title{Create formatted table with SIR results}
\usage{
create_parameter_table_SIR(
  sir_results_path,
  sig_dig = 3,
  interval_bounds = c(0.025, 0.975)
)
}
\arguments{
\item{sir_results_path}{character of length 1. Path to sir_results.csv file}

\item{sig_dig}{integer of length 1. Number of significant digits to be
reported in the table. Defaults to 3}

\item{interval_bounds}{numeric vector of length 2. Boundaries of the CI.
Default to c(0.025, 0.975)}
}
\value{
list of 2 tibbles
}
\description{
The point is to automatically create a table with parameter values and their
CI calculated by SIR. Returns a list of tibbles of length 2, the first
element being the sir_results.csv content, the second element being the
formatted table. NB : Residual error parameters are given as variances
}
\examples{
\dontrun{create_parameter_table_SIR("sir_results.csv")}
}
