% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_updated_modfiles.R
\name{create_updated_modfiles}
\alias{create_updated_modfiles}
\title{Create multiple NONMEM model files with updated initial estimates}
\usage{
create_updated_modfiles(
  dir_input_files = "",
  input_model_files,
  dir_output_files = "",
  output_model_files,
  overwrite = TRUE
)
}
\arguments{
\item{dir_input_files}{character. Contains path to directory containing the
input model files. Defaults to "".}

\item{input_model_files}{character vector of length 1 or n equal to the
number of output files. Contains name(s) of model(s) to be copied with updated
parameter initial estimates.}

\item{dir_output_files}{character. Contains path to directory containing the
output model files. Defaults to "".}

\item{output_model_files}{character vector of length n. Contains name(s) of
new model file(s).}

\item{overwrite}{boolean of length 1. If TRUE and output_model_files refers
to an already existing file, the existing file will be overwritten.
Defaults to TRUE.}
}
\value{
New .mod files created
}
\description{
The point of this function is to enable creation of multiple NONMEM model
files, with updated initial estimates when possible. It is possible to copy a
single model file multiple times or to create multiple output model files
from multiple input model files.
}
\examples{
\dontrun{create_updated_modfiles("run1.mod",c("run2.mod","run3.mod"))}
}
