% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_parameters_and_replace_model.R
\name{fix_parameters_and_replace_model}
\alias{fix_parameters_and_replace_model}
\title{Fix parameters and overwrite the model with the updated fixed parameters}
\usage{
fix_parameters_and_replace_model(
  dir_input_files,
  input_model_files,
  parameters_to_fix
)
}
\arguments{
\item{dir_input_files}{character of length 1. Contains path to directory containing the
input model files. Defaults to "".}

\item{input_model_files}{character vector of length 1 or n equal to the
number of input files. Contains name(s) of model(s) to be changed with some
fixed parameters}

\item{parameters_to_fix}{character of length 1 or p equal to the number of
parameters to be fixed. Contains the names of the parameters to be fixed.}
}
\value{
New .mod files created
}
\description{
What the title says. A model file and a character vector of parameters to be
fixed has to be provided
}
\examples{
#' \dontrun{fix_parameters_and_replace_model("../NONMEM/","run1.mod",c("KG","EMAX"))}
}
