% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfix_parameters_and_replace_model.R
\name{unfix_parameters_and_replace_model}
\alias{unfix_parameters_and_replace_model}
\title{Unfix parameters and overwrite the model with the updated unfixed parameters}
\usage{
unfix_parameters_and_replace_model(
  dir_input_files,
  input_model_files,
  parameters_to_unfix
)
}
\arguments{
\item{dir_input_files}{character of length 1 or n equal to the
number of input files. Contains path to directory containing the
input model files. Defaults to "".}

\item{input_model_files}{character vector of length 1 or n equal to the
number of input files. Contains name(s) of model(s) to be changed with some
unfixed parameters}

\item{parameters_to_unfix}{character of length 1 or p equal to the number of
parameters to be unfixed. Contains the names of the parameters to be unfixed.}
}
\value{
New .mod files created
}
\description{
What the title says. A model file and a character vector of parameters to be
unfixed has to be provided
}
\examples{
#' \dontrun{unfix_parameters_and_replace_model("../NONMEM/","run1.mod",c("KG","EMAX"))}
}
